api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
mangum==0.17.0
